Added -playerViewPreferredInitialLoadingView:
I've added this feature because sometimes it takes a lot of time before YouTube iframe player become ready, especially in bad connection situations. During that time users are not notified the view is actually loading or not. This can be handy.

Perhaps I should add a code to remove self.initialLoadingView when
-webView:didFailLoadWithError: as well to prevent the loading view from being displayed forever.