Remove initialLoadingView when webView is error
This will only be called when webview is completely failed to load iframe, which will be caused by no internet connection (or maybe when google is dead). Other errors will be handled by iframe API and in that case iframe is surely loaded and posts onReady: before onError: so no problem at all.