Set bundleId of app as the baseURL for the webView. (#371)
* Set bundleId of app as the baseURL for the webView.

Lazy load the originURL and always set it to the bundleId of the app. Prefix it with http:// if not the webView will not load correctly.

* Overwrite Youtube's iFrame origin attribute

We always want the origin attribute from the iFrame API (https://developers.google.com/youtube/player_parameters#origin) to be consistent with the webView.baseURL. This means we ignore what a user might set as origin in the playerVars.