Compare request.URL and originURL using lowercase Strings.
Otherwise if an app contains any Uppercasing in its name, we were encountering an issue in which request.URL.host and originURL.host were using different casing even tho the strings were the same and that was making the isEqual comparison fail.

Didn't encounter this issue before because the sample app we test against does not contain Uppercase in its name.